Parallelism stuff.

+++ runtime/thread_pool.h
@@ -2,10 +2,11 @@
 #define _THREAD_POOL_H_
 
 #include <pthread.h>
+#include <stdint.h>
 
 #include "job.h"
 
-typedef void (*work_function_t)(int, int, void*, int*);
+typedef void (*work_function_t)(int, int, void*, int64_t*);
 
 typedef enum {
   THREAD_RUN = 0,
@@ -24,7 +25,7 @@ typedef struct {
   int                notify_when_done;
   work_function_t    work_function;
   void              *args;
-  int               *tile_sizes;
+  int64_t           *tile_sizes;
   int                iters_done;
   unsigned long long timestamp;
 } worker_data_t;
@@ -43,7 +44,7 @@ typedef struct {
 thread_pool_t *create_thread_pool(int max_threads);
 void launch_job(thread_pool_t *thread_pool,
                 work_function_t *work_functions, void **args, job_t *job,
-                int **tile_sizes, int reset_tps);
+                int64_t **tile_sizes, int reset_tps);
 void pause_job(thread_pool_t *thread_pool);
 int job_finished(thread_pool_t *thread_pool);
 int get_iters_done(thread_pool_t *thread_pool);
